(#32 #33) Unify instructor, member, user services
User service now handles all member, instructor and
generic users.  Moved some of the createIfNotFound
methods of InitialDataSetup in the config to user service.

This was part of the reason behind #32.

I found it easier this was to keep all user type services
in one interface (though the implementation is a little
lengthy, there is possible refactoring here but not before
the course end).

Modified the default data to use the correct member.

Minor fixes in templates with moving to a base controller
common ModelAndView approach.

Fixes #32 Fixes #33