(#20 #34) Members (and all users) can register for class
Fixes #20.  Members can now register and deregister
(revoke their registration) of any class in the future.

#34 is quasi-fixed, or bandaged in the sense that I changed
the onetomany mapping in Clazz to what I think is right
(and doesn't stack trace in strange combos).  The
weak-entity composite key is no longer a real composite
key.  I'm not sure how to map that in hibernate.