(#45 #35 #25) Class edits
Most of this update focuses on class edits.

I had a lot of trouble getting mockmvc to work with
the throwable catch blocks (I didn't get it working).

So a lot of tests go un-tested.  I should do it at the
service layer, but thought I had the mockmvc working.

It currently works with manual tests.

Small changes to postgres-schema for updates (i.e. we
don't check overlaps with the record being updated).

validation works for class edit updates.